# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TradeUpdate API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Submit signup request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "200":
          description: Signup submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthLoginData"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Account not approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthLoginStatus"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
        success:
          type: boolean
...
    SignupInput:
      type: object
      required: [username, email, country, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 20
        email:
          type: string
          format: email
        country:
          type: string
        password:
          type: string
          minLength: 8
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthLoginData:
      type: object
      required: [role, username]
      properties:
        role:
          type: string
        username:
          type: string
        mustChangePassword:
          type: boolean
        hasProfile:
          type: boolean
        hasToken:
          type: boolean
    AuthLoginStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
    CurrentUser:
      type: object
      required: [id, username, role, status]
      properties:
        id:
          type: string
        username:
          type: string
        role:
          type: string
        status:
          type: string
        email:
          type: string
        tradingProfile:
          type: ["string", "null"]
        strategyId:
          type: ["string", "null"]
        accountBalance:
          type: number
        peakBalance:
          type: number
        dailyStartBalance:
          type: number
        autoCompoundEnabled:
          type: boolean
        adaptiveIntelligenceEnabled:
          type: boolean
        copyTradingEnabled:
          type: boolean
        mustChangePassword:
          type: boolean
        hasDerivToken:
          type: boolean
        tradingMode:
          type: ["string", "null"]
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
...